Add armor stand shopkeepers
Changes to how we handle entity interactions: Armor stands require handling
the PlayerInteractAtEntity event. We previously canceled but otherwise
ignored this event, because for most mobs, it is called in addition to the
usual PlayerInteractEntity event (we only want to handle each interaction
once).
Now, instead of ignoring all PlayerInteractAtEntity events, which breaks
armor stand interactions, we track for each player their last interacted
entity and ignore subsequent interaction events for the same entity in a
short time frame.