NamedArgument no longer supports nested fallbacks
Instead, the inner argument is expected to immediately resolve to either a
result or a parsing error.

If the argument name prefix was successfully parsed and the inner argument
fails to parse, we then abort the whole command argument parsing by
throwing an ArgumentRejectedException: This ensures that the named argument
input does not end up beng parsed by an earlier command argument fallback,
resulting in a less relevant argument binding or parsing error.